# Extractor dfs performance