@kylesayrs 2024-07-05
Related: https://github.com/onnx/onnx/pull/6213