@jeanniefinks 2025-05-10
Per the main [README announcement](https://github.com/neuralmagic/deepsparse/blob/main/README.md), DeepSparse is being deprecated by June 2, 2025. Closing the PR as work has been suspended; thank you for the inputs and support!